Changed placeholder `ARRAY` to `a`
I wrote `ARRAY` in ALLCAPS to mark it as a placeholder, not because an actual variable should have this format. However, bash beginners often think variables in bash should be written in allcaps and could see this as a confirmation.
Allcaps variables are by convention environment variables. Using allcaps names for all variables is not only hard to read, but also risky since there might be accidental name collisions with environment variables and special variables like `PWD`.